Refactor websocket tests and update backend expectations
