+++ tests/test_ws_client.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import asyncio
+import copy
 import json
 import logging
 import types
@@ -15,8 +16,6 @@ _install_stubs()
 
 import custom_components.termoweb.ws_client as ws_core
 
-ws_client_legacy = ws_core
-
 
 def _load_ws_client(
     *,
@@ -24,7 +23,7 @@ def _load_ws_client(
     ws_connect_results: Iterable[Any] | None = None,
 ):
     _install_stubs()
-    module = ws_client_legacy
+    module = ws_core
     testing = module.aiohttp.testing
     defaults = getattr(testing, "_defaults", None)
     if defaults is not None:
@@ -2028,3 +2027,360 @@ def test_mark_event_unique_paths() -> None:
     finally:
         module._LOGGER.setLevel(old_level)
     assert client._stats.last_paths == ["/a", "/b", "/c", "/d", "/e"]
+
+
+def test_detect_protocol_uses_hint_and_base() -> None:
+    module = _load_ws_client()
+
+    loop = types.SimpleNamespace(create_task=lambda *_args, **_kwargs: None)
+    hass = types.SimpleNamespace(loop=loop, data={module.DOMAIN: {"entry": {}}})
+    coordinator = types.SimpleNamespace()
+    client = module.TermoWebSocketClient(
+        hass,
+        entry_id="entry",
+        dev_id="dev",
+        api_client=types.SimpleNamespace(api_base="https://api-tevolve.example", _session=None),
+        coordinator=coordinator,
+    )
+
+    assert client._detect_protocol() == "engineio2"
+
+    client._protocol_hint = "socketio09"
+    assert client._detect_protocol() == "socketio09"
+
+
+def test_engineio_ws_client_flow(
+    monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture
+) -> None:
+    module = _load_ws_client()
+
+    dispatcher_calls: list[tuple[str, dict[str, object]]] = []
+
+    def fake_dispatcher(hass, signal: str, payload: dict[str, object]) -> None:
+        dispatcher_calls.append((signal, payload))
+
+    monkeypatch.setattr(module, "async_dispatcher_send", fake_dispatcher)
+    monkeypatch.setattr(ws_core, "async_dispatcher_send", fake_dispatcher)
+
+    async def _run() -> None:
+        loop = asyncio.get_running_loop()
+        caplog.set_level(logging.DEBUG, logger=module.__name__)
+
+        node_updates: list[tuple[dict[str, Any], list[Any]]] = []
+        energy_updates: list[Any] = []
+
+        class RecordingCoordinator:
+            def update_nodes(self, nodes: dict[str, Any], inventory: list[Any]) -> None:
+                node_updates.append((copy.deepcopy(nodes), list(inventory)))
+
+        class FakeEnergyCoordinator:
+            def update_addresses(self, addrs: Iterable[str] | dict[str, Iterable[str]]) -> None:
+                if isinstance(addrs, dict):
+                    energy_updates.append({k: list(v) for k, v in addrs.items()})
+                else:
+                    energy_updates.append(list(addrs))
+
+        hass = types.SimpleNamespace(
+            loop=loop,
+            data={
+                module.DOMAIN: {
+                    "entry": {
+                        "ws_state": {},
+                        "energy_coordinator": FakeEnergyCoordinator(),
+                    }
+                }
+            },
+        )
+        coordinator = RecordingCoordinator()
+
+        class FakeClient:
+            api_base = "https://api-tevolve.termoweb.net/"
+
+            async def _authed_headers(self) -> dict[str, str]:
+                return {"Authorization": "Bearer tok"}
+
+        client = module.TermoWebSocketClient(
+            hass,
+            entry_id="entry",
+            dev_id="dev",
+            api_client=FakeClient(),
+            coordinator=coordinator,
+            protocol="engineio2",
+        )
+
+        assert client._detect_protocol() == "engineio2"
+        client._protocol = "engineio2"
+        client._update_status("starting")
+        client._update_status("connecting")
+        client._update_status("connected")
+
+        handshake_payload = {
+            "devs": [{"id": "dev"}],
+            "permissions": {"dev": ["read"]},
+        }
+        client._on_frame(
+            json.dumps({"event": "dev_handshake", "data": handshake_payload})
+        )
+        await asyncio.sleep(0)
+        handshake_payload["devs"][0]["id"] = "mutated"
+        assert client._handshake is not None
+        assert client._handshake["devs"][0]["id"] == "dev"
+
+        initial_update = {
+            "nodes": {
+                "htr": {"status": {"01": {"temp": 20}}},
+                "acm": {"status": {"A1": {"mode": "eco"}}},
+            }
+        }
+        client._on_frame(json.dumps({"event": "update", "data": initial_update}))
+        await asyncio.sleep(0)
+
+        dev_data_payload = {
+            "nodes": {
+                "htr": {"status": {"01": {"temp": 20}}},
+                "acm": {"status": {"A1": {"mode": "eco"}}},
+                "raw": {"meta": {"foo": "bar"}},
+            }
+        }
+        client._on_frame(json.dumps({"event": "dev_data", "data": dev_data_payload}))
+        await asyncio.sleep(0)
+
+        incremental_update = {
+            "nodes": {
+                "htr": {"status": {"02": {"temp": 21}}},
+                "acm": {"status": {"A2": {"mode": "boost"}}},
+                "raw": {"meta": {"foo": "bar", "extra": True}},
+                "metrics": 3,
+            }
+        }
+        client._on_frame(json.dumps({"event": "update", "data": incremental_update}))
+        await asyncio.sleep(0)
+
+        client._on_frame(json.dumps({"event": "update", "data": None}))
+        client._on_frame(json.dumps({"event": "unknown", "data": {}}))
+        client._on_frame(json.dumps("literal"))
+        client._on_frame("not-json")
+        await asyncio.sleep(0)
+
+        await client.stop()
+        await asyncio.sleep(0)
+
+        status_signal = module.signal_ws_status("entry")
+        data_signal = module.signal_ws_data("entry")
+
+        status_updates = [
+            payload["status"]
+            for signal, payload in dispatcher_calls
+            if signal == status_signal
+        ]
+        assert status_updates[0] == "starting"
+        assert "healthy" in status_updates
+        assert status_updates[-1] == "stopped"
+
+        data_payloads = [
+            payload for signal, payload in dispatcher_calls if signal == data_signal
+        ]
+        assert len(data_payloads) == 3
+        assert data_payloads[0]["nodes"] == initial_update["nodes"]
+        assert data_payloads[1]["nodes"] == dev_data_payload["nodes"]
+        assert data_payloads[2]["nodes"]["htr"]["status"] == {
+            "01": {"temp": 20},
+            "02": {"temp": 21},
+        }
+        assert data_payloads[2]["nodes"]["acm"]["status"] == {
+            "A1": {"mode": "eco"},
+            "A2": {"mode": "boost"},
+        }
+        assert data_payloads[2]["nodes"]["raw"] == {
+            "meta": {"foo": "bar", "extra": True},
+        }
+        assert len(node_updates) == len(energy_updates) == 3
+        assert node_updates[0][0] == initial_update["nodes"]
+        assert node_updates[1][0] == dev_data_payload["nodes"]
+        assert node_updates[2][0]["htr"]["status"]["02"]["temp"] == 21
+        assert node_updates[2][0]["acm"]["status"]["A2"]["mode"] == "boost"
+        assert all(not update for update in energy_updates)
+        assert client._nodes["nodes_by_type"]["htr"]["status"]["02"]["temp"] == 21
+        assert client._nodes["nodes_by_type"]["acm"]["status"]["A2"]["mode"] == "boost"
+        assert client._healthy_since is not None
+        assert client._status == "stopped"
+        assert "unknown node types" not in caplog.text
+
+    asyncio.run(_run())
+
+
+def test_engineio_logs_unknown_types(
+    monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture
+) -> None:
+    module = _load_ws_client()
+
+    dispatcher_calls: list[tuple[str, dict[str, object]]] = []
+
+    def fake_dispatcher(hass, signal: str, payload: dict[str, object]) -> None:
+        dispatcher_calls.append((signal, payload))
+
+    monkeypatch.setattr(module, "async_dispatcher_send", fake_dispatcher)
+    monkeypatch.setattr(ws_core, "async_dispatcher_send", fake_dispatcher)
+
+    async def _run() -> None:
+        loop = asyncio.get_running_loop()
+        caplog.set_level(logging.DEBUG, logger=module.__name__)
+
+        node_updates: list[tuple[dict[str, Any], list[Any]]] = []
+        energy_updates: list[Any] = []
+
+        class RecordingCoordinator:
+            def update_nodes(self, nodes: dict[str, Any], inventory: list[Any]) -> None:
+                node_updates.append((copy.deepcopy(nodes), list(inventory)))
+
+        class FakeEnergyCoordinator:
+            def update_addresses(self, addrs: Iterable[str] | dict[str, Iterable[str]]) -> None:
+                if isinstance(addrs, dict):
+                    energy_updates.append({k: list(v) for k, v in addrs.items()})
+                else:
+                    energy_updates.append(list(addrs))
+
+        hass = types.SimpleNamespace(
+            loop=loop,
+            data={
+                module.DOMAIN: {
+                    "entry": {
+                        "ws_state": {},
+                        "energy_coordinator": FakeEnergyCoordinator(),
+                    }
+                }
+            },
+        )
+        coordinator = RecordingCoordinator()
+
+        class FakeClient:
+            api_base = "https://api-tevolve.termoweb.net/"
+
+            async def _authed_headers(self) -> dict[str, str]:
+                return {"Authorization": "Bearer tok"}
+
+        client = module.TermoWebSocketClient(
+            hass,
+            entry_id="entry",
+            dev_id="dev",
+            api_client=FakeClient(),
+            coordinator=coordinator,
+            protocol="engineio2",
+        )
+
+        client._protocol = "engineio2"
+        payload = {
+            "nodes": {
+                "nodes": [
+                    {"addr": "X", "type": "gizmo"},
+                    {"addr": "01", "type": "htr"},
+                ]
+            }
+        }
+        client._handle_dev_data(payload)
+        await asyncio.sleep(0)
+
+        assert any(
+            "unknown node types in inventory: gizmo" in rec.message
+            for rec in caplog.records
+        )
+        assert energy_updates == [{"gizmo": ["X"], "htr": ["01"]}]
+        assert node_updates
+
+    asyncio.run(_run())
+
+
+def test_engineio_stop_handles_cancelled_task(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    module = _load_ws_client()
+
+    dispatcher_calls: list[tuple[str, dict[str, object]]] = []
+
+    def fake_dispatcher(hass, signal: str, payload: dict[str, object]) -> None:
+        dispatcher_calls.append((signal, payload))
+
+    monkeypatch.setattr(module, "async_dispatcher_send", fake_dispatcher)
+    monkeypatch.setattr(ws_core, "async_dispatcher_send", fake_dispatcher)
+
+    async def _run() -> None:
+        loop = asyncio.get_running_loop()
+        hass = types.SimpleNamespace(loop=loop, data={})
+
+        class FakeClient:
+            api_base = "https://api-tevolve.termoweb.net/"
+
+            async def _authed_headers(self) -> dict[str, str]:
+                return {"Authorization": "Bearer token"}
+
+        client = module.TermoWebSocketClient(
+            hass,
+            entry_id="entry",
+            dev_id="dev",
+            api_client=FakeClient(),
+            coordinator=types.SimpleNamespace(),
+            protocol="engineio2",
+        )
+
+        async def hanging_runner(self: module.TermoWebSocketClient) -> None:
+            self._protocol = "engineio2"
+            self._update_status("connecting")
+            try:
+                await asyncio.Future()
+            finally:
+                self._update_status("stopped")
+
+        client._runner = types.MethodType(hanging_runner, client)
+
+        task = client.start()
+        await asyncio.sleep(0)
+        assert not task.done()
+
+        task.cancel()
+        await asyncio.sleep(0)
+
+        await client.stop()
+        await asyncio.sleep(0)
+
+        assert client._task is None
+        status_signal = module.signal_ws_status("entry")
+        status_updates = [
+            payload["status"]
+            for signal, payload in dispatcher_calls
+            if signal == status_signal
+        ]
+        assert status_updates[-1] == "stopped"
+
+    asyncio.run(_run())
+
+
+def test_engineio_ws_url_requires_token() -> None:
+    module = _load_ws_client()
+
+    async def _run() -> None:
+        loop = asyncio.get_running_loop()
+        hass = types.SimpleNamespace(
+            loop=loop,
+            data={module.DOMAIN: {"entry": {"ws_state": {}}}},
+        )
+        coordinator = types.SimpleNamespace()
+
+        class NoTokenClient:
+            api_base = "https://api-tevolve.termoweb.net/"
+
+            async def _authed_headers(self) -> dict[str, str]:
+                return {}
+
+        client = module.TermoWebSocketClient(
+            hass,
+            entry_id="entry",
+            dev_id="dev",
+            api_client=NoTokenClient(),
+            coordinator=coordinator,
+            protocol="engineio2",
+        )
+
+        with pytest.raises(RuntimeError):
+            await client.ws_url()
+
+    asyncio.run(_run())
